refactor: split local_comfy.py into resolve / mutate / nested-exec modules
